Move handle_opcode_result to execution_loop.rs
