+++ crates/vm/levm/src/vm.rs
@@ -48,7 +48,7 @@ pub struct Substate {
 ///   - Substate
 ///   - Gas Refunds
 ///   - Transient Storage
-struct Backup {
+pub struct Backup {
     cache: CacheDB,
     substate: Substate,
     refunded_gas: u64,
@@ -264,14 +264,14 @@ impl VM {
         }
     }
 
-    fn restore_state(&mut self, backup: Backup) {
+    pub fn restore_state(&mut self, backup: Backup) {
         self.cache = backup.cache;
         self.accrued_substate = backup.substate;
         self.env.refunded_gas = backup.refunded_gas;
         self.env.transient_storage = backup.transient_storage;
     }
 
-    fn is_create(&self) -> bool {
+    pub fn is_create(&self) -> bool {
         matches!(self.tx_kind, TxKind::Create)
     }
 
@@ -1093,82 +1093,6 @@ impl VM {
         }
 
         op_result
-    }
-
-    fn handle_opcode_result(
-        &mut self,
-        reason: ResultReason,
-        current_call_frame: &mut CallFrame,
-        backup: Backup,
-    ) -> Result<TransactionReport, VMError> {
-        self.call_frames.push(current_call_frame.clone());
-        // On successful create check output validity
-        if (self.is_create() && current_call_frame.depth == 0)
-            || current_call_frame.create_op_called
-        {
-            let contract_code = std::mem::take(&mut current_call_frame.output);
-            let code_length = contract_code.len();
-
-            let code_length_u64: u64 = code_length
-                .try_into()
-                .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-
-            let code_deposit_cost: u64 =
-                code_length_u64
-                    .checked_mul(CODE_DEPOSIT_COST)
-                    .ok_or(VMError::Internal(
-                        InternalError::ArithmeticOperationOverflow,
-                    ))?;
-
-            // Revert
-            // If the first byte of code is 0xef
-            // If the code_length > MAX_CODE_SIZE
-            // If current_consumed_gas + code_deposit_cost > gas_limit
-            let validate_create = if code_length > MAX_CODE_SIZE {
-                Err(VMError::ContractOutputTooBig)
-            } else if contract_code.first().unwrap_or(&0) == &INVALID_CONTRACT_PREFIX {
-                Err(VMError::InvalidContractPrefix)
-            } else if self
-                .increase_consumed_gas(current_call_frame, code_deposit_cost)
-                .is_err()
-            {
-                Err(VMError::OutOfGas(OutOfGasError::MaxGasLimitExceeded))
-            } else {
-                Ok(current_call_frame.to)
-            };
-
-            match validate_create {
-                Ok(new_address) => {
-                    // Set bytecode to new account if success
-                    update_account_bytecode(&mut self.cache, &self.db, new_address, contract_code)?;
-                }
-                Err(error) => {
-                    // Revert if error
-                    current_call_frame.gas_used = current_call_frame.gas_limit;
-                    self.restore_state(backup);
-
-                    return Ok(TransactionReport {
-                        result: TxResult::Revert(error),
-                        new_state: HashMap::default(),
-                        gas_used: current_call_frame.gas_used,
-                        gas_refunded: self.env.refunded_gas,
-                        output: std::mem::take(&mut current_call_frame.output),
-                        logs: std::mem::take(&mut current_call_frame.logs),
-                        created_address: None,
-                    });
-                }
-            }
-        }
-
-        return Ok(TransactionReport {
-            result: TxResult::Success,
-            new_state: HashMap::default(),
-            gas_used: current_call_frame.gas_used,
-            gas_refunded: self.env.refunded_gas,
-            output: std::mem::take(&mut current_call_frame.output),
-            logs: std::mem::take(&mut current_call_frame.logs),
-            created_address: None,
-        });
     }
 
     fn handle_opcode_error(
